requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
Django==2.1.3
djangorestframework==3.9.0
factory-boy==2.11.1
Faker==1.0.0
flake8==3.6.0
mccabe==0.6.1
more-itertools==4.3.0
mypy==0.641
mypy-extensions==0.4.1
pluggy==0.8.0
py==1.7.0
pycodestyle==2.4.0
pyflakes==2.0.0
pytest==4.0.0
pytest-django==3.4.4
python-dateutil==2.7.5
pytz==2018.7
rope==0.11.0
six==1.11.0
text-unidecode==1.2
typed-ast==1.1.0
